packages/stonksfeed/pyproject.toml
[project]
name = "stonksfeed"
version = "0.1.0"
description = "Stock news aggregator - RSS feeds and forum scrapers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dateutil>=2.8.0",
    "pytz>=2024.1",
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "moto[dynamodb]>=5.0.0",
    "responses>=0.25.0",
]

[project.scripts]
stonksfeed = "stonksfeed.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stonksfeed"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
